+++ frontend/src/lib/components/apps/components/display/dbtable/AppDbExplorer.svelte
@@ -64,6 +64,10 @@
 			lastTable = undefined
 		}
 
+		clearColumnDefs()
+	}
+
+	function clearColumnDefs() {
 		const gridItem = findGridItem($app, id)
 
 		if (!gridItem) {
@@ -445,6 +449,11 @@
 
 		state = undefined
 
+		// If in the mean time the table has changed, we don't want to update the columnDefs
+		if (lastTable !== table) {
+			return
+		}
+
 		//@ts-ignore
 		gridItem.data.configuration.columnDefs = { value: ncols, type: 'static', loading: false }
 		gridItem.data = gridItem.data
@@ -669,6 +678,11 @@
 					on:update={onUpdate}
 					on:delete={onDelete}
 					allowColumnDefsActions={false}
+					on:recompute={() => {
+						lastTable = undefined
+						clearColumnDefs()
+						listColumnsIfAvailable()
+					}}
 					{actions}
 				/>
 			{/key}