+++ frontend/src/lib/components/apps/editor/settingsPanel/decisionTree/DecisionTreePreview.svelte
@@ -222,13 +222,18 @@
 			}
 
 			case 'delete': {
-				const graphhNodeIndex = nodes.findIndex((node) => node.id == graphNode?.id)
+				const graphNodeIndex = nodes.findIndex((node) => node.id == graphNode?.id)
 
-				if (graphhNodeIndex > -1) {
-					deleteSubgrid(graphhNodeIndex)
+				if (graphNodeIndex > -1) {
+					deleteSubgrid(graphNodeIndex)
 				}
 
 				nodes = removeNode(nodes, graphNode)
+
+				$debuggingComponents = Object.fromEntries(
+					Object.entries($debuggingComponents).filter(([key]) => key !== component.id)
+				)
+
 				break
 			}
 			case 'addBranch': {