+++ frontend/src/lib/components/runs/JobLoader.svelte
@@ -106,12 +106,15 @@
 
 	async function fetchJobs(
 		startedBefore: string | undefined,
-		startedAfter: string | undefined
+		startedAfter: string | undefined,
+		startedAfterCompletedJobs: string | undefined
 	): Promise<Job[]> {
+		console.log('fetching jobs', startedAfter, startedBefore)
 		return JobService.listJobs({
 			workspace: $workspaceStore!,
 			createdOrStartedBefore: startedBefore,
 			createdOrStartedAfter: startedAfter,
+			createdOrStartedAfterCompletedJobs: startedAfterCompletedJobs,
 			schedulePath,
 			scriptPathExact: path === null || path === '' ? undefined : path,
 			createdBy: user === null || user === '' ? undefined : user,
@@ -137,7 +140,8 @@
 	async function fetchExtendedJobs(
 		concurrencyKey: string | null,
 		startedBefore: string | undefined,
-		startedAfter: string | undefined
+		startedAfter: string | undefined,
+		startedAfterCompletedJobs: string | undefined
 	): Promise<ExtendedJobs> {
 		return ConcurrencyGroupsService.listExtendedJobs({
 			rowLimit: 1000,
@@ -145,6 +149,7 @@
 			workspace: $workspaceStore!,
 			createdOrStartedBefore: startedBefore,
 			createdOrStartedAfter: startedAfter,
+			createdOrStartedAfterCompletedJobs: startedAfterCompletedJobs,
 			schedulePath,
 			scriptPathExact: path === null || path === '' ? undefined : path,
 			createdBy: user === null || user === '' ? undefined : user,
@@ -192,7 +197,7 @@
 		loading = true
 		try {
 			if (concurrencyKey == null || concurrencyKey === '') {
-				let newJobs = await fetchJobs(maxTs, undefined)
+				let newJobs = await fetchJobs(maxTs, undefined, minTs)
 				extendedJobs = { jobs: newJobs, obscured_jobs: [] } as ExtendedJobs
 
 				// Filter on minTs here and not in the backend
@@ -200,7 +205,7 @@
 				jobs = sortMinDate(minTs, newJobs)
 				externalJobs = []
 			} else {
-				extendedJobs = await fetchExtendedJobs(concurrencyKey, maxTs, undefined)
+				extendedJobs = await fetchExtendedJobs(concurrencyKey, maxTs, undefined, minTs)
 				const newJobs = extendedJobs.jobs
 				const newExternalJobs = extendedJobs.obscured_jobs
 
@@ -285,10 +290,10 @@
 					loading = true
 					let newJobs: Job[]
 					if (concurrencyKey == null || concurrencyKey === '') {
-						newJobs = await fetchJobs(maxTs, minTs ?? ts)
+						newJobs = await fetchJobs(maxTs, minTs ?? ts, undefined)
 					} else {
 						// Obscured jobs have no ids, so we have to do the full request
-						extendedJobs = await fetchExtendedJobs(concurrencyKey, maxTs, undefined)
+						extendedJobs = await fetchExtendedJobs(concurrencyKey, maxTs, undefined, minTs ?? ts)
 						externalJobs = computeExternalJobs(extendedJobs.obscured_jobs)
 
 						// Filter on minTs here and not in the backend